Minor tweaks
1. Starting flows + progress
2. Better complete flow shape
3. Don't show lives in tutorial
4. Bugfix: Music from wrong (default) theme at launch

+++ Assets/Scripts/Data/PointData.cs
@@ -5,7 +5,7 @@ namespace Dev.Krk.MemoryFlow.Data
     [Serializable]
     public class PointData
     {
-        public int X;
-        public int Y;
+        public float X;
+        public float Y;
     }
 }
